:sparkles: ADD new server
