+++ package.json
@@ -30,8 +30,12 @@
     "@fastify/view": "^7.4.1",
     "dotenv": "^16.0.3",
     "eta": "^2.0.1",
+    "express": "^4.18.2",
+    "express-handlebars": "^7.0.6",
+    "express-rate-limit": "^6.7.0",
     "fastify": "^4.15.0",
     "node-appwrite": "^8.2.0",
+    "node-fetch": "^2.6.9",
     "openai": "^3.2.1"
   },
   "devDependencies": {
