+++ src/scss/settings/_font.scss
@@ -1,6 +1,6 @@
 @font-face {
   font-family: Zephyr;
-  src: url("/public/assets/ZephyrRegular.woff");
+  src: url("/assets/ZephyrRegular.woff");
   font-weight: 400;
   font-style: normal;
 }
