+++ views/layouts/main.hbs
@@ -1,0 +1,11 @@
+<!DOCTYPE html>
+<html lang="nl">
+{{> head }}
+<body>
+{{> header }}
+    <main>
+        {{{body}}}
+    </main>
+    {{> scripts }}
+</body>
+</html>
