+++ views/partials/head.hbs
@@ -1,0 +1,12 @@
+<head>
+  <meta charset="UTF-8" />
+  <meta http-equiv="X-UA-Compatible" content="IE=edge" />
+  <meta
+    name="viewport"
+    content="width=device-width, initial-scale=1.0, maximum-scale=1.0, user-scalable=0"
+  />
+  <title>{{page.title}} | Dichter</title>
+  <meta name="description" content="Dichter app" />
+  <link rel="stylesheet" href="/css/all.css" />
+  <script src="/scripts/index.js" type="module"></script>
+</head>
